pyproject.toml
[project]
name = "opencatalog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.38.6",
    "ipykernel>=6.29.5",
    "jinja2>=3.1.6",
    "pandas>=2.2.3",
    "pyarrow>=20.0.0",
    "pyiceberg>=0.9.1",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "snowflake-snowpark-python>=1.32.0",
]
